backend/requirements.txt
fastapi>=0.109.1
uvicorn[standard]==0.27.0
python-multipart>=0.0.22
pydantic==2.5.3
chromadb==0.4.22
sentence-transformers==2.3.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy<2.0
